+++ packages/communication/ezrassor_joy_translator/source/ezrassor_joy_translator/joy_translator.py
@@ -4,7 +4,7 @@
 Written by Harrison Black.
 """
 import rospy
-from std_msgs.msg import Int8, Int16
+from std_msgs.msg import Int8, Int32
 from sensor_msgs.msg import Joy
 import time
 
@@ -12,7 +12,7 @@ NODE = "joy_translator"
 TOPIC = "joy"
 
 # Global so topic initiated at start of node
-publisher = rospy.Publisher('/ezrassor/requests', Int16, queue_size = 10)
+publisher = rospy.Publisher('/ezrassor/requests', Int32, queue_size = 10)
 tank_turn = False
 toggle_time = 0
